refactor: use infinite query options factory
