+++ src/components/FeedSelector.tsx
@@ -1,5 +1,5 @@
 import { useAuth } from '../lib/bluesky/hooks/useAuth';
-import { useFeeds } from '../lib/bluesky/hooks/useFeeds';
+import { getFeeds, useFeeds } from '../lib/bluesky/hooks/useFeeds';
 import { usePreferences } from '../lib/bluesky/hooks/usePreferences';
 import { useSettings } from '../hooks/useSetting';
 import * as Ariakit from '@ariakit/react';
@@ -12,29 +12,7 @@ export const FeedSelector = ({ columnNumber = 1 }: { columnNumber: number }) => 
   const { t } = useTranslation('app');
   const { isAuthenticated } = useAuth();
   const { data: preferences } = usePreferences();
-  const savedFeedsPrefV2 = isAuthenticated
-    ? preferences?.find((item) => item.$type === 'app.bsky.actor.defs#savedFeedsPrefV2')
-    : null;
-  const feeds = (
-    savedFeedsPrefV2?.items as
-      | (
-          | {
-              type: 'feed';
-              value: `at://${string}`;
-              pinned: boolean;
-              id: string;
-            }
-          | {
-              type: 'timeline';
-              value: string;
-              pinned: boolean;
-              id: string;
-            }
-        )[]
-      | undefined
-  )
-    ?.filter((item) => item.type === 'feed')
-    ?.map((item) => item.value) ?? ['at://did:plc:z72i7hdynmk6r22z27h6tvur/app.bsky.feed.generator/whats-hot'];
+  const feeds = getFeeds({ isAuthenticated, preferences });
   const { isLoading, error, data } = useFeeds({ feeds });
   const selectedFeed = columns[columnNumber] ?? feeds[0];
