+++ src/lib/bluesky/hooks/useFeeds.ts
@@ -1,5 +1,6 @@
 import { queryOptions, useQuery } from '@tanstack/react-query';
 import { type BlueskyState, useBlueskyStore } from '../store';
+import type { usePreferences } from './usePreferences';
 
 type FeedsQueryOptions = Pick<BlueskyState, 'agent'> & {
   feeds: string[];
@@ -25,3 +26,36 @@ export function useFeeds({ feeds }: { feeds: string[] }) {
 
   return useQuery(feedsQueryOptions({ agent, feeds }));
 }
+
+interface FeedSelectorParams {
+  isAuthenticated: boolean;
+  preferences: ReturnType<typeof usePreferences>['data'];
+}
+
+export const getFeeds = ({ isAuthenticated, preferences }: FeedSelectorParams) => {
+  const savedFeedsPrefV2 = isAuthenticated
+    ? preferences?.find((item) => item.$type === 'app.bsky.actor.defs#savedFeedsPrefV2')
+    : null;
+  const feeds = (
+    savedFeedsPrefV2?.items as
+      | (
+          | {
+              type: 'feed';
+              value: `at://${string}`;
+              pinned: boolean;
+              id: string;
+            }
+          | {
+              type: 'timeline';
+              value: string;
+              pinned: boolean;
+              id: string;
+            }
+        )[]
+      | undefined
+  )
+    ?.filter((item) => item.type === 'feed')
+    ?.map((item) => item.value) ?? ['at://did:plc:z72i7hdynmk6r22z27h6tvur/app.bsky.feed.generator/whats-hot'];
+
+  return feeds;
+};
