+++ src/lib/bluesky/hooks/useTimeline.ts
@@ -1,7 +1,8 @@
-import { useInfiniteQuery } from '@tanstack/react-query';
-import { useBlueskyStore } from '../store';
+import { infiniteQueryOptions, useInfiniteQuery } from '@tanstack/react-query';
+import { type BlueskyState, useBlueskyStore } from '../store';
 import { usePreferences } from './usePreferences';
 import { BskyPost } from '../types';
+import { getFeeds } from './useFeeds';
 
 type Timeline = {
   feed: {
@@ -11,37 +12,17 @@ type Timeline = {
   cursor: string;
 };
 
-export function useTimeline(selectedFeed: string) {
-  const { agent, isAuthenticated } = useBlueskyStore();
-  const preferences = usePreferences();
-  const savedFeedsPrefV2 = isAuthenticated
-    ? preferences.data?.find((item) => item.$type === 'app.bsky.actor.defs#savedFeedsPrefV2')
-    : null;
-  const feeds = (
-    savedFeedsPrefV2?.items as
-      | (
-          | {
-              type: 'feed';
-              value: `at://${string}`;
-              pinned: boolean;
-              id: string;
-            }
-          | {
-              type: 'timeline';
-              value: string;
-              pinned: boolean;
-              id: string;
-            }
-        )[]
-      | undefined
-  )
-    ?.filter((item) => item.type === 'feed')
-    ?.map((item) => item.value) ?? ['at://did:plc:z72i7hdynmk6r22z27h6tvur/app.bsky.feed.generator/whats-hot'];
-  const feed = feeds.find((feed) => feed === selectedFeed) ?? feeds[0];
+type TimelineQueryOptions = Pick<BlueskyState, 'agent' | 'isAuthenticated'> & {
+  selectedFeed: string;
+  preferences: ReturnType<typeof usePreferences>['data'];
+};
 
-  return useInfiniteQuery<Timeline>({
+export const timelineQueryOptions = ({ agent, isAuthenticated, selectedFeed, preferences }: TimelineQueryOptions) => {
+  const feeds = getFeeds({ isAuthenticated, preferences });
+  const feed = feeds.find((feed) => feed === selectedFeed) ?? feeds[0];
+  return infiniteQueryOptions({
     queryKey: ['timeline', { feed, isAuthenticated }],
-    queryFn: async ({ pageParam }) => {
+    queryFn: async ({ pageParam: cursor }) => {
       if (!agent) {
         throw new Error('Not authenticated');
       }
@@ -54,7 +35,6 @@ export function useTimeline(selectedFeed: string) {
       // }
 
       // authenticated
-      const cursor = pageParam as string | undefined;
       const response = await agent.api.app.bsky.feed.getFeed({
         feed,
         cursor,
@@ -63,8 +43,15 @@ export function useTimeline(selectedFeed: string) {
       return response.data as Timeline;
     },
     getNextPageParam: (lastPage) => lastPage.cursor,
-    initialPageParam: undefined,
+    initialPageParam: undefined as string | undefined,
     enabled: !!agent && feed !== undefined,
     retry: 1,
   });
+};
+
+export function useTimeline(selectedFeed: string) {
+  const { agent, isAuthenticated } = useBlueskyStore();
+  const { data: preferences } = usePreferences();
+
+  return useInfiniteQuery(timelineQueryOptions({ agent, isAuthenticated, selectedFeed, preferences }));
 }
